lkh_service/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
scikit-learn
numpy
